# Codechange: Build station and depot vehicle lists from shared order lists.