@PeterN 2024-01-05
> I'm seeing four times essentially the same loop, with only small differences in conditionals at different places. It looks to me like this can be written in a single templated function, with some small functions for the differentiating logic. Like whether to filter on company or not, what type of order is allowed, and how to add to a collection.
> 
> That would split the internals of the (complicated) loops from the actual important bits for each of the instances. It's bit similar to something like [`std::copy_if`](https://en.cppreference.com/w/cpp/algorithm/copy), which has an `UnaryPredicate` to separate the `if` logic from the looping logic.

I had a play with this and ended up with the function like this.

```
template <class VehiclePredicate, class OrderPredicate, class VehicleFunc>
void FindVehiclesWithOrder(VehiclePredicate vp, OrderPredicate op, VehicleFunc vf) {
...
}
```

This seems to be at least about 50% slower. Not sure if I'm doing it wrong, or if it's because it needs to capture variables.