@glx22 2024-01-05
> And here's a good thing. Making locals out of `ScriptObject::GetCompany()` and `ScriptCompanyMode::IsDeity()` made another performance difference:

Maybe do the same for `ScriptVehicleList` too.