@PeterN 2024-01-05
Performance difference since #11686:

dbg: [misc] [old] 207814 us [avg: 2078.1 us]
dbg: [misc] [new] 17636 us [avg: 176.4 us]
dbg: [misc] [old] 193357 us [avg: 1933.6 us]
dbg: [misc] [new] 16664 us [avg: 166.6 us]
dbg: [misc] [old] 203591 us [avg: 2035.9 us]
dbg: [misc] [new] 17441 us [avg: 174.4 us]
dbg: [misc] [old] 192353 us [avg: 1923.5 us]
dbg: [misc] [new] 16598 us [avg: 166.0 us]
dbg: [misc] [old] 202319 us [avg: 2023.2 us]
dbg: [misc] [new] 17576 us [avg: 175.8 us]

But please double check I haven't got the conditions wrong :)