@PeterN 2024-01-05
Splitting the deity check seems to save on average about 20µs (so about 10%) in the stationlist case.